feat: Add multi location view to ViewDetailsSheet.dart
